fix(react-ui-editor): Use a single extension for inline markup decorations
