--- packages/ui/react-ui-editor/src/extensions/markdown/heading.ts
@@ -1,59 +1,0 @@
-//
-// Copyright 2023 DXOS.org
-//
-
-import { syntaxTree } from '@codemirror/language';
-import { RangeSetBuilder } from '@codemirror/state';
-import { Decoration, type DecorationSet, type EditorView, ViewPlugin, type ViewUpdate } from '@codemirror/view';
-
-const buildDecorations = (view: EditorView): DecorationSet => {
-  const builder = new RangeSetBuilder<Decoration>();
-  const { state } = view;
-  const cursor = state.selection.main.head;
-
-  for (const { from, to } of view.visibleRanges) {
-    syntaxTree(state).iterate({
-      enter: (node) => {
-        switch (node.name) {
-          case 'ATXHeading1':
-          case 'ATXHeading2':
-          case 'ATXHeading3':
-          case 'ATXHeading4':
-          case 'ATXHeading5':
-          case 'ATXHeading6': {
-            const mark = node.node.getChild('HeaderMark');
-            if (mark) {
-              if (!view.hasFocus || view.state.readOnly || cursor < node.from || cursor > node.to) {
-                builder.add(mark.from, mark.to + 1, Decoration.replace({}));
-              }
-            }
-          }
-        }
-      },
-      from,
-      to,
-    });
-  }
-
-  return builder.finish();
-};
-
-export const heading = () => {
-  return [
-    ViewPlugin.fromClass(
-      class {
-        decorations: DecorationSet;
-        constructor(view: EditorView) {
-          this.decorations = buildDecorations(view);
-        }
-
-        update(update: ViewUpdate) {
-          this.decorations = buildDecorations(update.view);
-        }
-      },
-      {
-        decorations: (value) => value.decorations,
-      },
-    ),
-  ];
-};
