+++ packages/ui/react-ui-editor/src/extensions/markdown/index.ts
@@ -3,12 +3,9 @@
 //
 
 export * from './bundle';
-export * from './code';
+export * from './decorate';
 export * from './formatting';
-export * from './heading';
 export * from './highlight';
-export * from './hr';
 export * from './image';
 export * from './link';
 export * from './table';
-export * from './tasklist';
